Separate feature config from abyss context
