+++ gradle.properties
@@ -1,3 +1,3 @@
 group=com.mineinabyss
 version=0.11
-idofrontVersion=0.19.0
+idofrontVersion=0.19.3
