+++ mineinabyss-core/src/main/kotlin/com/mineinabyss/mineinabyss/core/AbyssFeature.kt
@@ -1,16 +1,9 @@
 package com.mineinabyss.mineinabyss.core
 
-import com.mineinabyss.idofront.config.IdofrontConfig
-
 interface AbyssFeature {
     val dependsOn: Set<String> get() = setOf()
 
     fun MineInAbyssPlugin.enableFeature() {}
 
     fun MineInAbyssPlugin.disableFeature() {}
-}
-
-interface Configurable<T> {
-    val configManager: IdofrontConfig<T>
-    val config: T
 }
