+++ mineinabyss-core/src/main/kotlin/com/mineinabyss/mineinabyss/core/Configurable.kt
@@ -1,0 +1,8 @@
+package com.mineinabyss.mineinabyss.core
+
+import com.mineinabyss.idofront.config.IdofrontConfig
+
+interface Configurable<T> {
+    val configManager: IdofrontConfig<T>
+    val config: T get() = configManager.getOrLoad()
+}
