+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/features/configs/ConfigReloadsFeature.kt
@@ -3,19 +3,16 @@ package com.mineinabyss.features.configs
 import com.mineinabyss.idofront.commands.arguments.stringArg
 import com.mineinabyss.idofront.messaging.error
 import com.mineinabyss.idofront.messaging.success
-import com.mineinabyss.mineinabyss.core.AbyssFeature
-import com.mineinabyss.mineinabyss.core.Configurable
-import com.mineinabyss.mineinabyss.core.MineInAbyssPlugin
-import com.mineinabyss.mineinabyss.core.abyss
+import com.mineinabyss.mineinabyss.core.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
 
 @Serializable
-@SerialName("configs")
-class ConfigsFeature : AbyssFeature {
+@SerialName("config_reloads")
+class ConfigReloadsFeature : AbyssFeature {
     @Transient
-    private val configs = abyss.config.features.filterIsInstance<Configurable<*>>().associateBy { it.configManager.name }
+    private val configs = abyssFeatures.features.filterIsInstance<Configurable<*>>().associateBy { it.configManager.name }
 
     override fun MineInAbyssPlugin.enableFeature() {
         commands {
