+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/features/okibotravel/OkiboTravelListener.kt
@@ -4,6 +4,7 @@ import com.destroystokyo.paper.event.player.PlayerUseUnknownEntityEvent
 import com.github.shynixn.mccoroutine.bukkit.launch
 import com.mineinabyss.components.okibotravel.OkiboTraveler
 import com.mineinabyss.components.playerData
+import com.mineinabyss.features.helpers.di.Features.okiboLine
 import com.mineinabyss.geary.helpers.with
 import com.mineinabyss.geary.papermc.tracking.entities.toGeary
 import com.mineinabyss.idofront.messaging.error
