--- src/main/resources/config.yml
@@ -1,41 +1,0 @@
-# Example config:
-
-# Extensions
-x-particles-vomit: &vomit
-  count: 30
-  duration: 10s
-  iterations: 6
-  offset: 3s
-  particles: [ SLIME ]
-
-# Layer config uses DeeperWorld sections
-# effect names are each effect class' @SerialName
-features:
-  - !<all> {}
-layers:
-  - name: Orth
-    sub: City of the Great Pit
-    deathMessage: in the city of Orth
-    depth:
-      start: -1
-      end: 0
-    sections: [ section1 ]
-    blockBlacklist:
-      - LADDER
-  - name: Edge of the Abyss
-    sub: 0-1350 meters
-    deathMessage: on the Edge of the Abyss
-    depth:
-      start: 0
-      end: 1350
-    effects:
-      - !<potion>
-        strength: 2
-        duration: 3s
-        effects: [ HUNGER ]
-      - !<particles>
-        <<: *vomit
-    sections: [ section2 ]
-    blockBlacklist:
-      - LADDER
-hubSectionName: section1
